Add Dark theme

+++ src/modules/projects/ProjectsPage.vue
@@ -5,6 +5,7 @@ import { useCurrentUserStore } from '../../shared/stores/currentUser'
 import { useProjectsStore } from '../../shared/stores/projects'
 
 import ProjectCreateForm from './components/ProjectCreateForm.vue'
+import ProjectsList from './components/ProjectsList.vue'
 
 const currentUserStore = useCurrentUserStore()
 const projectsStore = useProjectsStore()
@@ -12,6 +13,11 @@ const projectsStore = useProjectsStore()
 const currentUserId = computed(() => currentUserStore.currentUserId)
 const loadingProjectsData = computed(() => projectsStore.loadingProjectsData)
 const projects = computed(() => projectsStore.projects)
+
+const currentProjectId = computed(() => {
+  if (currentProject.value) return currentProject.value.id
+  return 0
+})
 
 const mod = ref('create')
 const newProjectName = ref('')
@@ -35,9 +41,8 @@ const createProject = () => {
   projectsStore.addProject({ project: newProject })
 }
 
-const editProject = project => {
-  currentProject.value = project
-  newProjectName.value = project.name
+const editProject = () => {
+  newProjectName.value = currentProject.value.name
   mod.value = 'edit'
 }
 
@@ -56,28 +61,44 @@ const saveProject = () => {
 }
 
 const deleteProject = projectId => {
-  projectsStore.deleteProject({ projectId })
+  if (confirm('Удалить проект?')) {
+    projectsStore.deleteProject({ projectId })
+    currentProject.value = null
+  }
 }
 </script>
 
 <template>
   <div>
     <!-- <div>ProjectsPage</div> -->
+
     <div class="row">
-      <div class="col-6">
+      <div
+        class="my-vh100 col-sm-3 col-lg-2 d-none d-sm-flex flex-column border-end border-dark-subtle p-0 pb-2"
+      >
+        <ProjectsList
+          :projects="projects"
+          :currentProjectId="currentProjectId"
+          @set-current-project="setCurrentProject"
+          @edit-project="editProject"
+          @delete-project="deleteProject"
+        />
+      </div>
+
+      <div class="col-12 col-sm-9 col-lg-10">
         <ProjectCreateForm
           @save-project="saveProject"
           v-model="newProjectName"
           :mod="mod"
         />
-      </div>
-      <div class="col-6">
-        <pre>{{ currentProject }}</pre>
+
+        <div v-if="currentProject">
+          <pre>{{ currentProject }}</pre>
+        </div>
       </div>
     </div>
 
     <div>
-      <!-- <div>Total projects: {{ projectsStore.projectsLength }}</div> -->
       <div id="projects-list" class="mt-3">
         <div v-for="project in projects" :key="project.id" data-test="project">
           <button class="btn btn-sm" @click="setCurrentProject(project)">
